refactor(tests): reorganize test files into __tests__ directory
Move all test files from src root to dedicated __tests__ directory to
improve project structure and follow common testing conventions. This
separation makes the codebase more organized and aligns with standard
JavaScript/TypeScript project layouts for better maintainability.
