+++ src/__tests__/convert-to-qwen-chat-messages.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from "vitest"
-import { convertToQwenChatMessages } from "./convert-to-qwen-chat-messages"
+import { convertToQwenChatMessages } from "../convert-to-qwen-chat-messages"
 
 vi.stubEnv("DASHSCOPE_API_KEY", "test-api-key-123")
